Validate datasets based on meta record, not table
